requirements.txt
openai==1.51.0
python-dotenv==1.0.1
